API/requirements.txt
django
djangorestframework
environs
requests